5-app-infra/3-artifact-publish/docker/flex-templates/gcs_to_bq_deidentification/requirements.txt
apache-beam[gcp]==2.54.0  # Must match the version in `Dockerfile``.
tink[gcpkms]==1.10.0
google-cloud-secret-manager==2.20.2
google-cloud-dlp==3.12.2

